requirements.txt
asn1crypto==0.24.0
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
cryptography==2.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.1
django-toolbelt==0.0.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.8
googlemaps==3.0.2
grpcio==1.18.0
gunicorn==19.9.0
idna==2.8
Pillow==5.4.1
protobuf==3.6.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
static3==0.7.0
urllib3==1.24.1
